pyproject.toml
[project]
name = "mbase"
version = "0.1.0"
description = "A professional CLI for Metabase"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "appdirs>=1.4",
]
[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]
[project.scripts]
mbase = "mbase.main:app"
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/mbase"]
